backend/requirements.txt
Django==5.2.10
channels==4.1.0
daphne==4.1.2
websockets==12.0
djangorestframework==3.14.0
psycopg2-binary==2.9.9
channels-redis==4.2.0
redis==5.0.1
python-dotenv==1.2.1